Репозиторий каталогов. Реализована архитектура с возможностью настройки имплементации (свойство repository.mode): работа с памятью (repository.mode=memory) или работа с БД Postgres (repository.mode=postgres). Класс CatalogRepository преобразован в интерфейс.
