+++ src/main/resources/application.properties
@@ -1,2 +1,3 @@
 spring.application.name=shop
 server.servlet.context-path=/api
+repository.mode=postgres
